scrap: rename codec.rs to vpxcodec.rs
